Second commit: fixed after review.

+++ src/main/java/core/basesyntax/Circle.java
@@ -16,8 +16,8 @@ public class Circle extends Figure {
     }
 
     @Override
-    public int getArea() {
-        return (int)(Math.PI * radius * radius);
+    public double getArea() {
+        return Math.PI * radius * radius;
     }
 
     @Override
@@ -26,5 +26,3 @@ public class Circle extends Figure {
                 + radius + " units, color: " + getColor());
     }
 }
-
-
